% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AoristicSum-class}
\alias{AoristicSum-class}
\alias{.AoristicSum}
\title{Aoristic Sum}
\description{
An S4 class to represent an aoristic analysis results.
}
\section{Slots}{

\describe{
\item{\code{from}}{A \code{\link{numeric}} vector.}

\item{\code{to}}{A \code{\link{numeric}} vector.}

\item{\code{weights}}{A \code{\link{numeric}} vector.}

\item{\code{groups}}{A \code{\link{character}} vector.}

\item{\code{dates}}{A \code{\link{numeric}} vector.}

\item{\code{p}}{A \code{\link{numeric}} \code{\link{array}} giving the aorisitic probabilities.}

\item{\code{sum}}{A \code{\link{numeric}} \code{\link{matrix}} giving the aorisitic sums (one group per
columns).}
}}

\section{Subset}{

In the code snippets below, \code{x} is an \code{AoristicSum} object.
\describe{
\item{\code{x[[i]]}}{Extracts information from a slot selected by subscript \code{i}.
\code{i} is a length-one \code{\link{character}} vector. Returns the corresponding slot
values.}
}
}

\section{Coerce}{

In the code snippets below, \code{x} is an \code{AoristicSum} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{CountApportion-class}},
\code{\link{DateEvent-class}},
\code{\link{DateMCD-class}},
\code{\link{IncrementTest-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
