% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CountApportion-class}
\alias{CountApportion-class}
\alias{.CountApportion}
\title{Aoristic Sum}
\description{
An S4 class to represent an artifact apportioning results.
}
\section{Slots}{

\describe{
\item{\code{p}}{An \code{\link{array}} giving the probability of apportioning an artifact type
to a given period.}

\item{\code{a}}{An \code{\link{array}} giving the apportioning of artifact types per site and
per period.}
}}

\seealso{
Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{DateEvent-class}},
\code{\link{DateMCD-class}},
\code{\link{IncrementTest-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
