% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{data.frame}
\alias{data.frame}
\alias{as.data.frame,MeanDate-method}
\alias{as.data.frame,AoristicSum-method}
\alias{as.data.frame,IncrementTest-method}
\title{Coerce to a Data Frame}
\usage{
\S4method{as.data.frame}{MeanDate}(x, ..., calendar = get_calendar())

\S4method{as.data.frame}{AoristicSum}(x, ..., calendar = get_calendar())

\S4method{as.data.frame}{IncrementTest}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further parameters to be passed to \code{\link[=data.frame]{data.frame()}}.}

\item{calendar}{An \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{aion::calendar()}}). If \code{NULL}, \emph{rata die} are returned.}

\item{row.names, optional}{Currently not used.}
}
\value{
A \code{\link{data.frame}} with an extra \code{time} column giving the (decimal) years at
which the time series was sampled.
}
\description{
Coerce to a Data Frame
}
\seealso{
Other mutators: 
\code{\link{mutators}},
\code{\link{series}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
