% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DateMCD-class}
\alias{DateMCD-class}
\alias{.DateMCD}
\title{Mean Ceramic Date}
\description{
S4 classes to store the Mean Ceramic Date of archaeological assemblages.
}
\section{Slots}{

\describe{
\item{\code{counts}}{A \code{\link{numeric}} matrix of count data.}

\item{\code{dates}}{A \code{\link{numeric}} vector of dates.}

\item{\code{mcd}}{A \code{\link{numeric}} vector giving the mean ceramic dates.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{DateMCD} object.
\describe{
\item{\code{x[[i]]}}{Extracts information from a slot selected by subscript \code{i}.
\code{i} is a length-one \code{\link{character}} vector. Returns the corresponding slot
values.}
}
}

\section{Coerce}{

In the code snippets below, \code{x} is a \code{DateMCD} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
\code{\link[=date_mcd]{date_mcd()}}

Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{DateEvent-class}},
\code{\link{IncrementTest-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
