% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecate.R
\docType{methods}
\name{deprecate}
\alias{deprecate}
\alias{refine}
\alias{refine-method}
\alias{refine,AveragePermutationOrder-method}
\alias{refine,CA-method}
\alias{refine,BootstrapCA-method}
\title{Deprecated Methods}
\usage{
refine(object, ...)

\S4method{refine}{AveragePermutationOrder}(object, cutoff, margin = c(1, 2), axes = c(1, 2), n = 30)

\S4method{refine}{BootstrapCA}(object, cutoff, margin = 1, axes = c(1, 2))
}
\arguments{
\item{object}{A \linkS4class{PermutationOrder} object (typically returned by
\code{\link[=seriate_average]{seriate_average()}}).}

\item{...}{Further arguments to be passed to internal methods.}

\item{cutoff}{A function that takes a numeric vector as argument and returns
a single numeric value (see below).}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscripts which the
refinement will be applied over: \code{1} indicates rows, \code{2} indicates columns.}

\item{axes}{An \code{\link{integer}} vector giving the subscripts of the CA axes to be
used.}

\item{n}{A non-negative \code{\link{integer}} giving the number of bootstrap
replications.}
}
\description{
Deprecated Methods
}
\author{
N. Frerebeau
}
\keyword{internal}
