% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EventDate-class}
\alias{EventDate-class}
\alias{.EventDate}
\title{Date Model}
\description{
S4 classes to store the event and accumulation times of archaeological
assemblages.
}
\section{Slots}{

\describe{
\item{\code{contexts}}{A \eqn{m \times p}{m x p} \code{\link{integer}} \code{\link{matrix}} of count
data.}

\item{\code{dates}}{A length-\eqn{m} \code{\link{numeric}} vector of dates.}

\item{\code{model}}{A \link[stats:lm]{multiple linear model}: the Gaussian
multiple linear regression model fitted for event date estimation and
prediction.}

\item{\code{keep}}{An \code{\link{integer}} vector.}
}}

\seealso{
\code{\link[dimensio:CA-class]{dimensio::CA}}

Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}},
\code{\link{PermutationOrder-class}},
\code{\link{RateOfChange-class}},
\code{\link{RefinePermutationOrder-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
