% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/test_fit.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{test_fit}
\alias{test_fit-method}
\alias{test_fit,CountMatrix-method}
\title{Frequency Increment Test}
\usage{
test_fit(object, ...)

\S4method{test_fit}{CountMatrix}(object, dates)
}
\arguments{
\item{object}{A \link[arkhe:CountMatrix-class]{arkhe::CountMatrix} object.}

\item{...}{Currently not used.}

\item{dates}{A \code{\link{numeric}} vector of dates.}
}
\value{
An \linkS4class{IncrementTest} object.
}
\description{
Frequency Increment Test
}
\details{
The Frequency Increment Test (FIT) rejects neutrality if the distribution
of normalized variant frequency increments exhibits a mean that deviates
significantly from zero.
}
\examples{
data("merzbach", package = "folio")

## Coerce the merzbach dataset to a count matrix
## Keep only decoration types that have a maximum frequency of at least 50
keep <- apply(X = merzbach, MARGIN = 2, FUN = function(x) max(x) >= 50)
counts <- as_count(merzbach[, keep])

## Group by phase
## We use the row names as time coordinates (roman numerals)
dates <- as.numeric(utils::as.roman(rownames(counts)))

## Plot abundance vs time
plot_time(counts, dates, facet = TRUE)

## Frequency Increment Test
freq <- test_fit(counts, dates)

## Plot time vs abundance and highlight selection
plot_time(freq)
plot_time(freq, roll = TRUE, window = 5)
}
\references{
Feder, A. F., Kryazhimskiy, S. & Plotkin, J. B. (2014). Identifying
Signatures of Selection in Genetic Time Series. \emph{Genetics}, 196(2):
509-522. \doi{10.1534/genetics.113.158220}.
}
\seealso{
\code{\link[=plot_time]{plot_time()}}

Other chronological analysis: 
\code{\link{aoristic}},
\code{\link{apportion}()}
}
\author{
N. Frerebeau
}
\concept{chronological analysis}
