% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khronos-package.R
\docType{package}
\name{khronos-package}
\alias{khronos-package}
\alias{_PACKAGE}
\alias{khronos}
\title{khronos: Time Point Estimate and Analysis of Chronological Patterns}
\description{
A toolkit for absolute dating and analysis of chronological
    patterns. This package includes functions for chronological modeling
    and dating of archaeological assemblages from count data.
}
\details{
\tabular{ll}{
\strong{Package:} \tab khronos \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tesselle/khronos}
  \item Report bugs at \url{https://github.com/tesselle/khronos/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
