% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cohort_display.R
\name{mod_cohort_display_ui}
\alias{mod_cohort_display_ui}
\alias{mod_cohort_display_server}
\title{mod_cohort_display_ui and mod_cohort_display_server}
\usage{
mod_cohort_display_ui(id)

mod_cohort_display_server(input, output, session, specimens)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
