% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_drug_screening.R
\name{mod_drug_screening_ui}
\alias{mod_drug_screening_ui}
\alias{mod_drug_screening_server}
\title{mod_drug_screening_ui and mod_drug_screening_server}
\usage{
mod_drug_screening_ui(id)

mod_drug_screening_server(input, output, session, specimens)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
