% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore_page.R
\name{mod_explore_page_ui}
\alias{mod_explore_page_ui}
\alias{mod_explore_page_server}
\title{mod_explore_page_ui and mod_explore_page_server}
\usage{
mod_explore_page_ui(id)

mod_explore_page_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
