% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gene_var.R
\name{mod_gene_variant_ui}
\alias{mod_gene_variant_ui}
\alias{mod_gene_variant_server}
\title{mod_gene_variant_ui and mod_gene_variant_server}
\usage{
mod_gene_variant_ui(id)

mod_gene_variant_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
