% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_page.R
\name{mod_home_page_ui}
\alias{mod_home_page_ui}
\alias{mod_home_page_server}
\title{mod_home_page_ui and mod_home_page_server}
\usage{
mod_home_page_ui(id)

mod_home_page_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
