% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_immune_signatures.R
\name{mod_immune_signatures_ui}
\alias{mod_immune_signatures_ui}
\alias{mod_immune_signatures_server}
\title{mod_immune_signatures_ui and mod_immune_signatures_server}
\usage{
mod_immune_signatures_ui(id)

mod_immune_signatures_server(input, output, session, specimens)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
