% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_latent_variables.R
\name{mod_latent_variables_ui}
\alias{mod_latent_variables_ui}
\alias{mod_latent_variables_server}
\title{mod_latent_variables_ui and mod_latent_variables_server}
\usage{
mod_latent_variables_ui(id)

mod_latent_variables_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
