% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/event_model.R
\docType{methods}
\name{model_event}
\alias{model_event}
\alias{summary,EventDate-method}
\alias{summary,EventDate,missing-method}
\alias{coef,EventDate-method}
\alias{fitted,EventDate-method}
\alias{residuals,EventDate-method}
\alias{sigma,EventDate-method}
\alias{terms,EventDate-method}
\title{Extract Event Date Model Results}
\usage{
\S4method{summary}{EventDate}(object, ...)

\S4method{coef}{EventDate}(object, calendar = NULL, ...)

\S4method{fitted}{EventDate}(object, calendar = NULL, ...)

\S4method{residuals}{EventDate}(object, calendar = NULL, ...)

\S4method{sigma}{EventDate}(object, calendar = NULL, ...)

\S4method{terms}{EventDate}(x, ...)
}
\arguments{
\item{...}{Currently not used.}

\item{calendar}{An \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{aion::calendar()}}). If \code{NULL} (the default), \emph{rata die} are
returned.}

\item{x, object}{An \code{\linkS4class{EventDate}} object.}
}
\description{
\itemize{
\item \code{summary()} summarizes linear model fit.
\item \code{coef()} extracts model coefficients (see \code{\link[stats:coef]{stats::coef()}}).
\item \code{fitted()} extracts model fitted values (see \code{\link[stats:fitted.values]{stats::fitted()}}).
\item \code{residuals()} extracts model residuals (see \code{\link[stats:residuals]{stats::residuals()}}).
\item \code{sigma()} extracts the residual standard deviation (see \code{\link[stats:sigma]{stats::sigma()}}).
\item \code{terms()} extracts model terms (see \code{\link[stats:terms]{stats::terms()}}).
}
}
\examples{
## Data from Peeples and Schachner 2012
data("zuni", package = "folio")

## Assume that some assemblages are reliably dated (this is NOT a real example)
zuni_dates <- c(
  LZ0569 = 1097, LZ0279 = 1119, CS16 = 1328, LZ0066 = 1111,
  LZ0852 = 1216, LZ1209 = 1251, CS144 = 1262, LZ0563 = 1206,
  LZ0329 = 1076, LZ0005Q = 859, LZ0322 = 1109, LZ0067 = 863,
  LZ0578 = 1180, LZ0227 = 1104, LZ0610 = 1074
)

## Model the event and accumulation date for each assemblage
model <- event(zuni, zuni_dates, rank = 10)
plot(model, select = 1:10, event = TRUE, flip = TRUE)
}
\seealso{
Other event date tools: 
\code{\link{density_event}()},
\code{\link{event}()},
\code{\link{predict_event}()}
}
\author{
N. Frerebeau
}
\concept{event date tools}
