% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{get_dates,DateMCD-method}
\title{Get or Set Parts of an Object}
\usage{
\S4method{get_dates}{DateMCD}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{x}.}
}
\value{
\itemize{
\item \verb{set_*()} returns an object of the same sort as \code{x} with the new values
assigned.
\item \verb{get_*()} returns the part of \code{x}.
}
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\seealso{
Other mutators: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
