% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PermutationOrder-class}
\alias{PermutationOrder-class}
\alias{.PermutationOrder}
\alias{RankPermutationOrder-class}
\alias{.RankPermutationOrder}
\alias{AveragePermutationOrder-class}
\alias{.AveragePermutationOrder}
\title{Permutation Order}
\description{
S4 classes to represent a permutation order.
}
\section{Slots}{

\describe{
\item{\code{rows_order}}{An \code{\link{integer}} vector giving the rows permutation.}

\item{\code{columns_order}}{An \code{\link{integer}} vector giving the columns permutation.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{PermutationOrder} object.
\describe{
\item{\code{x[[i]]}}{Extract information from a slot selected by subscript \code{i}.
\code{i} is a length-one character vector.}
}
}

\seealso{
\code{\link[dimensio:CA-class]{dimensio::CA}}

Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
