% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RateOfChange-class}
\alias{RateOfChange-class}
\alias{.RateOfChange}
\title{Rate of Change}
\description{
An S4 class to represent rates of change from an aoristic analysis.
}
\section{Slots}{

\describe{
\item{\code{replicates}}{A non-negative \code{\link{integer}} giving the number of
replications.}

\item{\code{blocks}}{A \code{\link{character}} vector giving the time-blocks.}

\item{\code{groups}}{A \code{\link{character}} vector.}

\item{\code{rates}}{A \code{\link{numeric}} \code{\link{array}} giving the rates of change.}
}}

\section{Subset}{

In the code snippets below, \code{x} is an \code{RateOfChange} object.
\describe{
\item{\code{x[[i]]}}{Extracts information from a slot selected by subscript \code{i}.
\code{i} is a length-one \code{\link{character}} vector. Returns the corresponding slot
values.}
}
}

\seealso{
Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{DateEvent-class}},
\code{\link{DateMCD-class}},
\code{\link{IncrementTest-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
