% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RefinePermutationOrder-class}
\alias{RefinePermutationOrder-class}
\alias{.RefinePermutationOrder}
\alias{RefineCA-class}
\title{Partial Bootstrap CA}
\description{
An S4 class to store partial bootstrap correspondence analysis results.
}
\section{Slots}{

\describe{
\item{\code{length}}{A \code{\link{numeric}} vector giving the convex hull maximum
dimension length.}

\item{\code{cutoff}}{A length-one \code{\link{numeric}} vector giving the cutoff value for
samples selection.}

\item{\code{keep}}{An \code{\link{integer}} vector giving the subscript of the variables
to be kept.}
}}

\seealso{
Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}},
\code{\link{PermutationOrder-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
