% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/simulate.R
\docType{methods}
\name{simulate.MeanDate}
\alias{simulate.MeanDate}
\alias{simulate,MeanDate-method}
\title{Simulate Mean Ceramic Dates}
\usage{
\S4method{simulate}{MeanDate}(object, nsim = 1000, seed = NULL, ...)
}
\arguments{
\item{object}{A \code{\linkS4class{MeanDate}} object (typically returned by \code{\link[=mcd]{mcd()}}).}

\item{nsim}{A non-negative \code{\link{integer}} specifying the number of simulations.}

\item{seed}{An object specifying if and how the random number generator
should be initialized (see \code{\link[stats:simulate]{stats::simulate()}}).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{SimulationMeanDate}} object.
}
\description{
Simulates \link[=mcd]{MCD} from a multinomial distribution.
}
\details{
For each assemblage, a large number of new bootstrap replicates
is created, with the same sample size, by resampling the original
assemblage with replacement. MCDs are calculated for each replicates.
}
\examples{
## Data from Peeples and Schachner 2012
data("zuni", package = "folio")

## Set the start and end dates for each ceramic type
dates <- list(
  LINO = c(600, 875), KIAT = c(850, 950), RED = c(900, 1050),
  GALL = c(1025, 1125), ESC = c(1050, 1150), PUBW = c(1050, 1150),
  RES = c(1000, 1200), TULA = c(1175, 1300), PINE = c(1275, 1350),
  PUBR = c(1000, 1200), WING = c(1100, 1200), WIPO = c(1125, 1225),
  SJ = c(1200, 1300), LSJ = c(1250, 1300), SPR = c(1250, 1300),
  PINER = c(1275, 1325), HESH = c(1275, 1450), KWAK = c(1275, 1450)
)

## Calculate date midpoints
mid <- vapply(X = dates, FUN = mean, FUN.VALUE = numeric(1))

## Calculate MCD
(mc_dates <- mcd(zuni[100:125, ], dates = mid))

## Get MCD in years CE
time(mc_dates, calendar = CE())

## Plot
plot(mc_dates)

## Bootstrap resampling
boot <- bootstrap(mc_dates, n = 30)
head(boot)

## Jackknife resampling
jack <- jackknife(mc_dates)
head(jack)

## Simulation
sim <- simulate(mc_dates, nsim = 30)
plot(sim, interval = "range", pch = 16)
}
\seealso{
\code{\link[=mcd]{mcd()}}

Other resampling methods: 
\code{\link{bootstrap.EventDate}},
\code{\link{bootstrap.MeanDate}},
\code{\link{jackknife.EventDate}},
\code{\link{jackknife.MeanDate}}
}
\author{
N. Frerebeau
}
\concept{resampling methods}
