% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SimulationMeanDate-class}
\alias{SimulationMeanDate-class}
\alias{.SimulationMeanDate}
\title{Simulated Mean Date}
\description{
An S4 class to store the simulated weighted mean date (e.g. Mean Ceramic
Date) of archaeological assemblages.
}
\section{Slots}{

\describe{
\item{\code{dates}}{A length-\eqn{p} \code{\link{numeric}} vector giving the dates of the
(ceramic) types expressed in \emph{\link[aion:RataDie-class]{rata die}}.}

\item{\code{replications}}{A \code{numeric} \code{\link{matrix}} giving the replications.}

\item{\code{seed}}{A \code{\link{numeric}} vector giving the value of \link{.Random.seed} before the
simulation was started; otherwise it is the user specified value with a
\code{kind} attribute  with value \verb{as.list(}\code{\link[=RNGkind]{RNGkind()}}\verb{)}.}
}}

\note{
Dates are internally stored as \emph{\link[aion:RataDie-class]{rata die}}.
This class inherits from \code{\linkS4class{MeanDate}}.
}
\seealso{
\code{\linkS4class{MeanDate}}

Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}},
\code{\link{PermutationOrder-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
