% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AoristicSum-class}
\alias{AoristicSum-class}
\alias{.AoristicSum}
\title{Aoristic Sum}
\description{
An S4 class to represent an aoristic analysis results.
}
\details{
Dates are internally stored as \emph{\link[aion:RataDie-class]{rata die}}.
}
\section{Slots}{

\describe{
\item{\code{weights}}{A \code{\link{numeric}} vector.}

\item{\code{breaks}}{A \code{\link{RataDie}} vector giving the date break between time-blocks.}

\item{\code{p}}{A \code{\link{numeric}} \code{\link{array}} giving the aorisitic probabilities.}

\item{\code{groups}}{A \code{\link{character}} vector.}
}}

\note{
This class inherits from \code{\link[aion:TimeSeries-class]{aion::TimeSeries}}.
}
\seealso{
Other classes: 
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}},
\code{\link{PermutationOrder-class}},
\code{\link{RateOfChange-class}},
\code{\link{RefinePermutationOrder-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
