% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AoristicSum-class}
\alias{AoristicSum-class}
\alias{.AoristicSum}
\title{Aoristic Sum}
\description{
An S4 class to represent an aoristic analysis results.
}
\section{Slots}{

\describe{
\item{\code{from}}{A \code{\link{numeric}} vector.}

\item{\code{to}}{A \code{\link{numeric}} vector.}

\item{\code{step}}{A length-one \code{\link{numeric}} vector giving the time-blocks width.}

\item{\code{weights}}{A \code{\link{numeric}} vector.}

\item{\code{breaks}}{A \code{\link{numeric}} vector giving the date break between time-blocks.}

\item{\code{blocks}}{A \code{\link{character}} vector giving the time-blocks.}

\item{\code{p}}{A \code{\link{numeric}} \code{\link{array}} giving the aorisitic probabilities.}

\item{\code{groups}}{A \code{\link{character}} vector.}
}}

\note{
This class inherits from base \code{\link{matrix}}.
}
\seealso{
Other classes: 
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
