% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CountApportion-class}
\alias{CountApportion-class}
\alias{.CountApportion}
\title{Count Apportioning}
\description{
An S4 class to represent an artifact apportioning results.
}
\section{Slots}{

\describe{
\item{\code{p}}{An \code{\link{array}} giving the probability of apportioning an artifact type
to a given period.}

\item{\code{apportion}}{An \code{\link{array}} giving the apportioning of artifact types per
site and per period.}

\item{\code{method}}{A \code{\link{character}} string specifying the distribution used for
apportioning (type popularity curve).}

\item{\code{from}}{A length-one \code{\link{numeric}} vector giving the beginning of the
period of interest (in years AD).}

\item{\code{to}}{A length-one \code{\link{numeric}} vector giving the end of the period of
interest (in years AD).}

\item{\code{step}}{A length-one \code{\link{integer}} vector giving the step size, i.e. the
width of each time step for apportioning (in years AD).}
}}

\seealso{
Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{DateEvent-class}},
\code{\link{DateMCD-class}},
\code{\link{IncrementTest-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
