% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DateMCD-class}
\alias{DateMCD-class}
\alias{.DateMCD}
\title{Mean Ceramic Date}
\description{
An S4 class to store the Mean Ceramic Date of archaeological assemblages.
}
\section{Slots}{

\describe{
\item{\code{dates_types}}{A \code{\link{numeric}} vector giving the dates of the ceramic types.}

\item{\code{dates_mcd}}{A \code{\link{numeric}} vector giving the mean ceramic dates.}
}}

\section{Coerce}{

In the code snippets below, \code{x} is a \code{DateMCD} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{DateEvent-class}},
\code{\link{IncrementTest-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
