% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kairos-deprecated.R
\name{kairos-deprecated}
\alias{kairos-deprecated}
\title{Deprecated Functions in kairos}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
