% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kairos-package.R
\docType{package}
\name{kairos-package}
\alias{kairos-package}
\alias{kairos}
\title{kairos: Analysis of Chronological Patterns from Archaeological Count Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolkit for absolute and relative dating and analysis of chronological patterns. This package includes functions for chronological modeling and dating of archaeological assemblages from count data. It provides methods for matrix seriation. It also allows to compute time point estimates and density estimates of the occupation and duration of an archaeological site.
}
\details{
\tabular{ll}{
\strong{Package:} \tab kairos \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 2.1.1 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\section{Package options}{

\code{kairos} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{kairos.progress}: a \code{\link{logical}} scalar. Should progress bars be
displayed?
\item \code{kairos.calendar}: a \code{\link[aion:TimeScale-class]{aion::TimeScale}} object (default calendar
for printing).
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/kairos/}
  \item \url{https://github.com/tesselle/kairos}
  \item Report bugs at \url{https://github.com/tesselle/kairos/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab  \cr
Ben Marwick \tab \emph{University of Washington, USA} \cr
Anne Philippe \tab \emph{Université de Nantes, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
