% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MeanDate-class}
\alias{MeanDate-class}
\alias{.MeanDate}
\title{Mean Date}
\arguments{
\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\description{
An S4 class to store the weighted mean date (e.g. Mean Ceramic Date) of
archaeological assemblages.
}
\section{Slots}{

\describe{
\item{\code{types}}{A length-\eqn{p} \code{\link{numeric}} vector giving the dates of the
(ceramic) types.}

\item{\code{weights}}{An \eqn{m \times p}{m x p} \code{\link{integer}} \code{\link{matrix}} giving the
weights used.}

\item{\code{simulation}}{A three columns \code{\link{numeric}} matrix giving the summary
statistics of the simulated dates (\code{mean} and \code{lower} and \code{upper} boundaries
of the confidence interval).}

\item{\code{replications}}{An \code{\link{integer}} giving the number of bootstrap
replications.}
}}

\note{
This class inherits from base \code{\link{numeric}}.
}
\section{Coerce}{

In the code snippets below, \code{x} is a \code{MeanDate} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{PermutationOrder-class}},
\code{\link{RateOfChange-class}},
\code{\link{RefinePermutationOrder-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
