% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{get_mcd}
\alias{get_mcd-method}
\alias{get_weights}
\alias{get_weights-method}
\alias{get_dates,AoristicSum-method}
\alias{get_dates,RateOfChange-method}
\alias{get_groups,AoristicSum-method}
\alias{get_mcd,DateMCD-method}
\alias{get_weights,AoristicSum-method}
\alias{get_weights,CountApportion-method}
\title{Get or Set Parts of an Object}
\usage{
get_mcd(x)

get_weights(x)

\S4method{get_dates}{AoristicSum}(x)

\S4method{get_dates}{RateOfChange}(x)

\S4method{get_groups}{AoristicSum}(x)

\S4method{get_mcd}{DateMCD}(x)

\S4method{get_weights}{AoristicSum}(x)

\S4method{get_weights}{CountApportion}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s).}
}
\value{
\itemize{
\item \verb{set_*()} returns an object of the same sort as \code{x} with the new values
assigned.
\item \verb{get_*()} returns the part of \code{x}.
}
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\seealso{
Other mutators: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
