% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{get_dates}
\alias{get_dates-method}
\alias{get_groups}
\alias{get_groups-method}
\alias{get_model}
\alias{get_model-method}
\alias{get_weights}
\alias{get_weights-method}
\alias{get_dates,AoristicSum-method}
\alias{get_dates,EventDate-method}
\alias{get_dates,RateOfChange-method}
\alias{get_groups,AoristicSum-method}
\alias{get_model,EventDate-method}
\alias{get_weights,AoristicSum-method}
\alias{get_weights,CountApportion-method}
\alias{get_weights,MeanDate-method}
\title{Get or Set Parts of an Object}
\usage{
get_dates(x)

get_groups(x)

get_model(x)

get_weights(x)

\S4method{get_dates}{AoristicSum}(x)

\S4method{get_dates}{EventDate}(x)

\S4method{get_dates}{RateOfChange}(x)

\S4method{get_groups}{AoristicSum}(x)

\S4method{get_model}{EventDate}(x)

\S4method{get_weights}{AoristicSum}(x)

\S4method{get_weights}{CountApportion}(x)

\S4method{get_weights}{MeanDate}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s).}
}
\value{
\itemize{
\item \verb{set_*()} returns an object of the same sort as \code{x} with the new values
assigned.
\item \verb{get_*()} returns the part of \code{x}.
}
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\author{
N. Frerebeau
}
\concept{mutators}
