% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{weights,AoristicSum-method}
\alias{weights,CountApportion-method}
\title{Get or Set Parts of an Object}
\usage{
\S4method{weights}{AoristicSum}(object, ...)

\S4method{weights}{CountApportion}(object, ...)
}
\arguments{
\item{object}{An object from which to get or set element(s).}

\item{...}{Currently not used.}
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\seealso{
Other mutators: 
\code{\link{data.frame}},
\code{\link{series}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
