% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/subset.R
\docType{class}
\name{PermutationOrder-class}
\alias{PermutationOrder-class}
\alias{.PermutationOrder}
\alias{RankPermutationOrder-class}
\alias{.RankPermutationOrder}
\alias{AveragePermutationOrder-class}
\alias{.AveragePermutationOrder}
\alias{[[,PermutationOrder,ANY,missing-method}
\title{Permutation Order}
\usage{
\S4method{[[}{PermutationOrder,ANY,missing}(x, i)
}
\arguments{
\item{x}{An object from which to extract element(s) or in which to replace
element(s).}

\item{i}{A \code{\link{character}} string specifying elements to extract.}
}
\description{
An S4 class to represent a permutation order.
}
\section{Functions}{
\itemize{
\item \code{x[[i}: Extracts information from a slot selected
by subscript i. i is a length-one character vector. Returns the
corresponding slot values.

}}
\section{Slots}{

\describe{
\item{\code{rows_order}}{An \code{\link{integer}} vector giving the rows permutation.}

\item{\code{columns_order}}{An \code{\link{integer}} vector giving the columns permutation.}

\item{\code{method}}{A \code{\link{character}} string indicating the seriation method used.}
}}

\seealso{
\code{\link[dimensio:CA-class]{dimensio::CA}}

Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}},
\code{\link{RateOfChange-class}},
\code{\link{RefinePermutationOrder-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
