% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mcd.R
\docType{methods}
\name{plot_mcd}
\alias{plot_mcd}
\alias{autoplot,MeanDate-method}
\alias{plot,MeanDate,missing-method}
\title{MCD Plot}
\usage{
\S4method{autoplot}{MeanDate}(object, ..., select = NULL, decreasing = TRUE)

\S4method{plot}{MeanDate,missing}(x, select = NULL, decreasing = TRUE, ...)
}
\arguments{
\item{object, x}{A \linkS4class{MeanDate} object.}

\item{...}{Currently not used.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector giving the selection of
the assemblage that are drawn.}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort be increasing or
decreasing?}
}
\value{
\itemize{
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
MCD Plot
}
\examples{
## Mean Ceramic Date
## Coerce the zuni dataset to an abundance (count) matrix
data("zuni", package = "folio")
counts <- as_count(zuni)

## Set the start and end dates for each ceramic type
dates <- list(
  LINO = c(600, 875), KIAT = c(850, 950), RED = c(900, 1050),
  GALL = c(1025, 1125), ESC = c(1050, 1150), PUBW = c(1050, 1150),
  RES = c(1000, 1200), TULA = c(1175, 1300), PINE = c(1275, 1350),
  PUBR = c(1000, 1200), WING = c(1100, 1200), WIPO = c(1125, 1225),
  SJ = c(1200, 1300), LSJ = c(1250, 1300), SPR = c(1250, 1300),
  PINER = c(1275, 1325), HESH = c(1275, 1450), KWAK = c(1275, 1450)
)

## Calculate date midpoints
mid <- vapply(X = dates, FUN = mean, FUN.VALUE = numeric(1))

## Calculate MCD
mc_dates <- mcd(counts, dates = mid)
head(mc_dates)

## Plot
plot(mc_dates, select = 100:125)

## Bootstrap resampling
boot <- bootstrap(mc_dates, n = 30)
head(boot)

## Jackknife resampling
jack <- jackknife(mc_dates)
head(jack)
}
\seealso{
\code{\link[=mcd]{mcd()}}

Other plotting methods: 
\code{\link{plot_aoristic}},
\code{\link{plot_event}},
\code{\link{plot_fit}},
\code{\link{plot_time}()}
}
\author{
N. Frerebeau
}
\concept{plotting methods}
