% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_time.R
\docType{methods}
\name{plot_time}
\alias{plot_time}
\alias{plot_time-method}
\alias{plot_time,CountMatrix,numeric-method}
\alias{plot_time,AoristicSum,ANY-method}
\alias{plot_time,AoristicSum,numeric-method}
\alias{plot_time,DateMCD,missing-method}
\alias{plot_time,IncrementTest,missing-method}
\title{Abundance vs Time Plot}
\usage{
plot_time(object, dates, ...)

\S4method{plot_time}{CountMatrix,numeric}(object, dates, facet = FALSE)

\S4method{plot_time}{AoristicSum,ANY}(object, facet = TRUE)

\S4method{plot_time}{DateMCD,missing}(object, facet = FALSE)

\S4method{plot_time}{IncrementTest,missing}(object, level = 0.95, roll = FALSE, window = 3)
}
\arguments{
\item{object}{A \link[arkhe:CountMatrix-class]{arkhe::CountMatrix} object.}

\item{dates}{A \code{\link{numeric}} vector of dates.}

\item{...}{Currently not used.}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels defined by
type/taxon be drawn?}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{roll}{A \code{\link{logical}} scalar: should each time series be
subsetted to look for episodes of selection?}

\item{window}{An odd \code{\link{integer}} giving the size of the rolling
window. Only used if \code{roll} is \code{TRUE}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Produces an abundance \emph{vs} time diagram.
}
\note{
Displaying FIT results on an abundance \emph{vs} time diagram is adapted from Ben
Marwick's original \href{https://github.com/benmarwick/signatselect/}{idea}.
}
\section{Detection of Selective Processes}{

Results of the frequency increment test can be displayed on an abundance
\emph{vs} time diagram aid in the detection and quantification of selective
processes in the archaeological record. If \code{roll} is \code{TRUE}, each time
series is subsetted according to \code{window} to see if episodes of selection
can be identified among decoration types that might not show overall
selection. If so, shading highlights the data points where
\code{\link[=test_fit]{test_fit()}} identifies selection.
}

\examples{
data("merzbach", package = "folio")

## Coerce the merzbach dataset to a count matrix
## Keep only decoration types that have a maximum frequency of at least 50
keep <- apply(X = merzbach, MARGIN = 2, FUN = function(x) max(x) >= 50)
counts <- as_count(merzbach[, keep])

## Group by phase
## We use the row names as time coordinates (roman numerals)
dates <- as.numeric(utils::as.roman(rownames(counts)))

## Plot abundance vs time
plot_time(counts, dates)
plot_time(counts, dates, facet = TRUE)
}
\seealso{
\code{\link[=date_mcd]{date_mcd()}}, \code{\link[=sum_aoristic]{sum_aoristic()}}, \code{\link[=test_fit]{test_fit()}}

Other plotting methods: 
\code{\link{plot_event}()}
}
\author{
N. Frerebeau
}
\concept{plotting methods}
