% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RateOfChange-class}
\alias{RateOfChange-class}
\alias{.RateOfChange}
\title{Rate of Change}
\description{
An S4 class to represent rates of change from an aoristic analysis.
}
\section{Slots}{

\describe{
\item{\code{replicates}}{A non-negative \code{\link{integer}} giving the number of
replications.}

\item{\code{breaks}}{A \code{\link{numeric}} vector giving the date break between time-blocks.}

\item{\code{groups}}{A \code{\link{character}} vector.}
}}

\note{
This class inherits from base \code{\link{array}}.
}
\seealso{
Other classes: 
\code{\link{AoristicSum-class}},
\code{\link{CountApportion-class}},
\code{\link{EventDate-class}},
\code{\link{IncrementTest-class}},
\code{\link{MeanDate-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
