% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_count}
\alias{as_long}
\alias{as_features}
\alias{get_samples}
\alias{set_samples<-}
\alias{get_groups}
\alias{set_groups<-}
\alias{get_dates}
\alias{set_dates<-}
\alias{get_tpq}
\alias{set_tpq<-}
\alias{get_taq}
\alias{set_taq<-}
\alias{CountMatrix}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{arkhe}{\code{\link[arkhe:CountMatrix-class]{CountMatrix}}, \code{\link[arkhe:coerce]{as_count}}, \code{\link[arkhe:coerce]{as_features}}, \code{\link[arkhe:coerce]{as_long}}, \code{\link[arkhe:mutator]{get_dates}}, \code{\link[arkhe:mutator]{get_groups}}, \code{\link[arkhe:mutator]{get_samples}}, \code{\link[arkhe:mutator]{get_taq}}, \code{\link[arkhe:mutator]{get_tpq}}, \code{\link[arkhe:mutator]{set_dates<-}}, \code{\link[arkhe:mutator]{set_groups<-}}, \code{\link[arkhe:mutator]{set_samples<-}}, \code{\link[arkhe:mutator]{set_taq<-}}, \code{\link[arkhe:mutator]{set_tpq<-}}}
}}

