% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{series}
\alias{series}
\alias{time,EventDate-method}
\alias{span,AoristicSum-method}
\title{Sampling Times}
\usage{
\S4method{time}{EventDate}(x, calendar = NULL)

\S4method{span}{AoristicSum}(x, calendar = NULL)
}
\arguments{
\item{x}{An \R object.}

\item{calendar}{An \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{aion::calendar()}}). If \code{NULL} (the default), \emph{rata die} are
returned.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the times at which a time series was sampled.
}
\seealso{
Other mutators: 
\code{\link{data.frame}},
\code{\link{mutators}},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
